requirements.txt
selenium
beautifulsoup4
python-dotenv == 0.19.2